renamed rootpage to searchpage
